pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-eating-storage-mysql"
version = "0.1.0"
description = "A MySQL storage solution for data eating operations"
authors = [{name = "Danilo Silva", email = "contact@danilocgsilva.me"}]
license = {text = "MIT"}
requires-python = ">=3.6"
dependencies = [
    "mysql-connector-python",
]

[project.urls]
Homepage = "https://github.com/danilocgsilva/data-eating-storage-mysql"